Create tribonacci
Tribonacci numbers are a sequence of numbers similar to Fibonacci numbers but with three initial values, typically 0, 0, and 1. The sequence starts with these three values, and each subsequent number is the sum of the previous three numbers. In mathematical notation, it can be defined as follows:

T(n) = T(n-1) + T(n-2) + T(n-3) for n > 2
where T(0) = 0, T(1) = 0, and T(2) = 1

Here are the first few terms of the Tribonacci sequence:
0, 0, 1, 1, 2, 4, 7, 13, 24, 44,…

Implement a recursive function (with memoization) in Python to generate the nth Tribonacci number.

Note: Make sure memoization technique has been used inside the tribonacci() function so that the program runs fast.