Create robots_on_grids
Robots on Grids
Let's consider a robot on an m x n grid of squares (that is, m rows and n columns). The robot starts
in the top-left square, and its target is at the bottom-right. The robot has two moves available to it
- it can move one square to the right, or one square down – so it must make a combination of
rightward and downward moves to reach its target. It follows that the robot will take (m + n – 2)
steps to reach its target.
Here are two such paths it could take:
Your task is to write code that counts the number of possible paths for the robot, given an m and n.
1. Write a function num_paths(m, n) that recursively computes the number of paths from the top-left
to bottom-right square of an m x n grid without memoization.
2. Write a function num_paths_memo(m, n) that does the same, but memoizes the smaller
subproblem solutions.
Expected Output
- The expected result for m=15, n=14 is 20058300.
- The elapsed time / runtime can vary from system to system.